Enemy creatures patrol system: (#11)
The enemy creature patrols the map. If it "sees" the player, the creature runs to the first position the players has when the creature saws him. If the creature fails his raimming, the creature waits several seconds and trys again (only if the player stil is in his hearing zone).
When the enemy collides with the player, the battle begins.